+++ library/src/main/java/it/gmariotti/cardslib/library/view/listener/SwipeDismissListViewTouchListener.java
@@ -36,6 +36,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+import it.gmariotti.cardslib.library.R;
 import it.gmariotti.cardslib.library.internal.Card;
 
 /**
@@ -104,6 +105,8 @@ public class SwipeDismissListViewTouchListener implements View.OnTouchListener {
     private View mDownView;
     private boolean mPaused;
 
+    private int swipeDistanceDivisor = 2;
+
     /**
      * The callback interface used by {@link SwipeDismissListViewTouchListener} to inform its client
      * about a successful dismissal of one or more list item positions.
@@ -141,6 +144,7 @@ public class SwipeDismissListViewTouchListener implements View.OnTouchListener {
                 android.R.integer.config_shortAnimTime);
         mListView = listView;
         mCallbacks = callbacks;
+        swipeDistanceDivisor =  listView.getContext().getResources().getInteger(R.integer.list_card_swipe_distance_divisor);
     }
 
     /**
@@ -217,10 +221,14 @@ public class SwipeDismissListViewTouchListener implements View.OnTouchListener {
                     mDownX = motionEvent.getRawX();
                     mDownY = motionEvent.getRawY();
                     mDownPosition = mListView.getPositionForView(mDownView);
-                    if (mCallbacks.canDismiss(mDownPosition,(Card) mListView.getAdapter().getItem(mDownPosition))) {
-                        mVelocityTracker = VelocityTracker.obtain();
-                        mVelocityTracker.addMovement(motionEvent);
-                    } else {
+                    if (mListView.getAdapter().getItem(mDownPosition) instanceof Card) {
+                        if (mCallbacks.canDismiss(mDownPosition, (Card) mListView.getAdapter().getItem(mDownPosition))) {
+                            mVelocityTracker = VelocityTracker.obtain();
+                            mVelocityTracker.addMovement(motionEvent);
+                        } else {
+                            mDownView = null;
+                        }
+                    }else{
                         mDownView = null;
                     }
                 }
@@ -242,7 +250,7 @@ public class SwipeDismissListViewTouchListener implements View.OnTouchListener {
                 float absVelocityY = Math.abs(mVelocityTracker.getYVelocity());
                 boolean dismiss = false;
                 boolean dismissRight = false;
-                if (Math.abs(deltaX) > mViewWidth / 2  && mSwiping) {
+                if (Math.abs(deltaX) > mViewWidth / swipeDistanceDivisor  && mSwiping) {
                     dismiss = true;
                     dismissRight = deltaX > 0;
                 } else if (mMinFlingVelocity <= absVelocityX && absVelocityX <= mMaxFlingVelocity